chore(app): typescript parcel react setup
