--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": ["react", "env"],
-  "plugins": [
-    ["transform-object-rest-spread", { "useBuiltIns": true }],
-    "transform-function-bind",
-    "transform-class-properties",
-    [
-      "module-resolver",
-      {
-        "root": ["./src"]
-      }
-    ]
-  ]
-}
