--- .eslintignore
@@ -1,4 +1,0 @@
-src/__helpers__/
-src/__mocks__/
-src/**/__tests__/
-jest.setup.js
