--- jest.setup.js
@@ -1,6 +1,0 @@
-import "raf/polyfill";
-
-import Enzyme from "enzyme";
-import Adapter from "enzyme-adapter-react-16";
-
-Enzyme.configure({ adapter: new Adapter() });
