+++ package.json
@@ -4,26 +4,18 @@
   "description": "Summary tool for neo-project, neo-ngd, CityOfZion, NewEconoLab and NeoResearch",
   "scripts": {
     "deploy": "node ghpages.js",
-    "build": "rimraf dist && parcel build src/index.html --no-cache --public-url /progress-summary",
-    "precommit": "yarn lint",
-    "lint": "eslint 'src/**/*.js*'",
-    "start": "parcel src/index.html",
-    "test": "jest --all --verbose --coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "test:local": "jest --all --verbose --coverage"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev:build": "parcel ./src/index.html --no-cache",
+    "build": "rimraf dist && parcel build ./src/index.html --no-cache --public-url /progress-summary",
+    "tslint": "npx tslint --config tslint.json --project tsconfig.json",
+    "prettier": "npx prettier --write src/**/*.{js,jsx,ts,tsx}",
+    "start": "npm run prettier && npm run tslint && npm run dev:build"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/be-neo/progress-summary.git"
   },
-  "keywords": [
-    "React",
-    "Parcel",
-    "JSS",
-    "JS",
-    "HTML",
-    "boilerplate",
-    "starter"
-  ],
+  "keywords": [],
   "author": "Jeroen Peeters <contact@jeroenpeeters.me>",
   "license": "MIT",
   "bugs": {
@@ -33,61 +25,29 @@
   "dependencies": {
     "antd": "3.8.1",
     "axios": "0.18.0",
-    "babel-polyfill": "6.26.0",
-    "jss": "9.8.7",
-    "jss-camel-case": "6.1.0",
-    "jss-global": "3.0.0",
-    "jss-vendor-prefixer": "8.0.1",
+    "babel-polyfill": "^6.26.0",
     "prop-types": "15.6.2",
     "react": "16.4.2",
     "react-dom": "16.4.2",
-    "react-jss": "8.6.1"
+    "react-ion-store": "^0.5.0",
+    "recompose": "^0.30.0"
   },
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-eslint": "8.2.6",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-class-properties": "6.24.1",
-    "babel-plugin-transform-function-bind": "6.22.0",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "babel-preset-env": "1.7.0",
-    "babel-preset-react": "6.24.1",
-    "coveralls": "3.0.2",
-    "enzyme": "3.4.3",
-    "enzyme-adapter-react-16": "1.2.0",
-    "enzyme-to-json": "3.3.4",
-    "eslint": "5.3.0",
-    "eslint-config-airbnb": "17.1.0",
-    "eslint-config-prettier": "2.9.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-jsx-a11y": "6.1.1",
-    "eslint-plugin-prettier": "2.6.2",
-    "eslint-plugin-react": "7.11.1",
+    "@babel/core": "^7.1.2",
+    "@types/react": "^16.4.14",
+    "@types/react-dom": "^16.0.8",
+    "@types/recompose": "^0.27.0",
+    "awesome-typescript-loader": "^5.2.1",
+    "babel-plugin-import": "^1.9.1",
     "gh-pages": "1.2.0",
-    "husky": "0.14.3",
-    "jest": "23.5.0",
-    "parcel-bundler": "1.9.7",
+    "less": "^2.7.3",
+    "parcel-bundler": "^1.10.1",
     "prettier": "1.14.2",
-    "raf": "3.4.0",
-    "react-test-renderer": "16.4.2",
-    "rimraf": "2.6.2"
-  },
-  "jest": {
-    "setupFiles": [
-      "./jest.setup.js"
-    ],
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}",
-      "!src/index.jsx",
-      "!src/__helpers__/TestPropTypes.helper.js",
-      "!**/node_modules/**"
-    ],
-    "moduleNameMapper": {
-      "^.+\\.(jpg|jpeg|png|gif|svg)$": "<rootDir>/src/__mocks__/file.mock.js"
-    },
-    "snapshotSerializers": [
-      "enzyme-to-json/serializer"
-    ]
+    "rimraf": "2.6.2",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.15.0",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.1.1"
   },
   "browserslist": [
     "> 1%",
