+++ src/App.less
@@ -1,0 +1,4 @@
+.ps__layout {
+  min-height: 100vh;
+  margin-left: 200px;
+}
