--- src/components/__tests__/SpinningLogo.spec.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { render } from "enzyme";
-import SpinningLogo from "./../SpinningLogo";
-
-describe("Components", () => {
-  describe("SpinningLogo", () => {
-    test("should render", () => {
-      const wrapper = render(<SpinningLogo />);
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-});
