+++ src/index.less
@@ -1,0 +1,1 @@
+@import "../node_modules/antd/dist/antd.less";
