+++ src/modules/Sider/styles.less
@@ -1,0 +1,15 @@
+.ps__sider__logo {
+  padding: 24px;
+
+  img {
+    width: 152px;
+  }
+}
+
+.ps__sider {
+  background: #fff;
+  overflow: auto;
+  height: 100vh;
+  position: fixed;
+  left: 0;
+}
