--- src/networking/consts.js
@@ -1,8 +1,0 @@
-export const GITHUB_API = "https://api.github.com";
-export const TOKEN_API = "https://wt-eb8e8a5788a32c0054649520e12aca04-0.sandbox.auth0-extend.com/nps-gh-tkn";
-export const ORGS = "/orgs";
-export const REPOS = "/repos";
-export const COMMITS = "/commits";
-export const PAGE = "&page=";
-export const SLASH = "/";
-export const ACCESS_TOKEN = token => `?access_token=${token}`;
