+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "strict": true,
+    "jsx": "react",
+    "lib": ["es5", "es6", "dom"],
+    "skipLibCheck": true
+  }
+}
