+++ tslint.json
@@ -1,0 +1,11 @@
+{
+  "defaultSeverity": "error",
+  "extends": ["tslint-react", "tslint-config-prettier"],
+  "jsRules": {},
+  "rules": {
+    "variable-name": false,
+    "function-name": false,
+    "jsx-no-lambda": false
+  },
+  "rulesDirectory": []
+}
